@@ -27,7 +27,7 @@
     }
   ],
   "require": {
-    "php": ">=7.4",
+    "php": "^8.2",
     "ext-json": "*"
   },
   "require-dev": {
