@@ -8,41 +8,41 @@ interface IResourceController
     /**
      * @return mixed
      */
-    public function index();
+    public function index(): mixed;
 
     /**
      * @param mixed $id
      * @return mixed
      */
-    public function show($id);
+    public function show(mixed $id): mixed;
 
     /**
      * @return mixed
      */
-    public function store();
+    public function store(): mixed;
 
     /**
      * @return mixed
      */
-    public function create();
+    public function create(): mixed;
 
     /**
      * View
      * @param mixed $id
      * @return mixed
      */
-    public function edit($id);
+    public function edit(mixed $id): mixed;
 
     /**
      * @param mixed $id
      * @return mixed
      */
-    public function update($id);
+    public function update(mixed $id): mixed;
 
     /**
      * @param mixed $id
      * @return mixed
      */
-    public function destroy($id);
+    public function destroy(mixed $id): mixed;
 
 }
