@@ -10,7 +10,7 @@ interface IClassLoader
      * @param string $class
      * @return object
      */
-    public function loadClass(string $class);
+    public function loadClass(string $class): object;
 
     /**
      * Called when loading class method
@@ -19,7 +19,7 @@ interface IClassLoader
      * @param array $parameters
      * @return mixed
      */
-    public function loadClassMethod($class, string $method, array $parameters);
+    public function loadClassMethod(object $class, string $method, array $parameters): mixed;
 
     /**
      * Called when loading method
@@ -28,6 +28,6 @@ interface IClassLoader
      * @param array $parameters
      * @return mixed
      */
-    public function loadClosure(Callable $closure, array $parameters);
+    public function loadClosure(Callable $closure, array $parameters): mixed;
 
 }
