@@ -2,7 +2,7 @@
 
 class CustomClassLoader implements \Pecee\SimpleRouter\ClassLoader\IClassLoader
 {
-    public function loadClass(string $class)
+    public function loadClass(string $class): object
     {
         return new DummyController();
     }
@@ -12,14 +12,19 @@ class CustomClassLoader implements \Pecee\SimpleRouter\ClassLoader\IClassLoader
      * @param object $class
      * @param string $method
      * @param array $parameters
-     * @return object
+     * @return mixed
      */
-    public function loadClassMethod($class, string $method, array $parameters)
+    public function loadClassMethod(object $class, string $method, array $parameters): mixed
     {
         return call_user_func_array([$class, $method], [true]);
     }
 
-    public function loadClosure(callable $closure, array $parameters)
+    /**
+     * @param callable $closure
+     * @param array $parameters
+     * @return mixed
+     */
+    public function loadClosure(callable $closure, array $parameters): mixed
     {
         return call_user_func_array($closure, [true]);
     }
