@@ -1,11 +1,16 @@
 <?php
 
-class TestRouter extends \Pecee\SimpleRouter\SimpleRouter
+use Pecee\Http\Exceptions\MalformedUrlException;
+use Pecee\Http\Url;
+use Pecee\SimpleRouter\Exceptions\NotFoundHttpException;
+use Pecee\SimpleRouter\SimpleRouter;
+
+class TestRouter extends SimpleRouter
 {
 
     public function __construct()
     {
-        static::request()->setHost('testhost.com');
+        static::request()->setHost('https://example.com');
     }
 
     public static function reset(): void
@@ -13,21 +18,29 @@ class TestRouter extends \Pecee\SimpleRouter\SimpleRouter
         static::$router = null;
     }
 
+    /**
+     * @throws \Pecee\SimpleRouter\Exceptions\HttpException
+     * @throws NotFoundHttpException
+     * @throws MalformedUrlException
+     */
     public static function debugNoReset(string $testUrl, string $testMethod = 'get'): void
     {
         $request = static::request();
 
-        $request->setUrl((new \Pecee\Http\Url($testUrl)));
+        $request->setUrl((new Url($testUrl)));
         $request->setMethod($testMethod);
 
         static::start();
     }
 
+    /**
+     * @throws Exception
+     */
     public static function debug(string $testUrl, string $testMethod = 'get', bool $reset = true): void
     {
         try {
             static::debugNoReset($testUrl, $testMethod);
-        } catch (\Exception $e) {
+        } catch (Exception $e) {
             static::$defaultNamespace = null;
             static::router()->reset();
             throw $e;
@@ -40,30 +53,32 @@ class TestRouter extends \Pecee\SimpleRouter\SimpleRouter
 
     }
 
+    /**
+     * @throws Exception
+     */
     public static function debugOutput(string $testUrl, string $testMethod = 'get', bool $reset = true): string
     {
-        $response = null;
-
         // Route request
         ob_start();
-        static::debug($testUrl, $testMethod, $reset);
-        $response = ob_get_clean();
 
+        static::debug($testUrl, $testMethod, $reset);
         // Return response
-        return $response;
+        return ob_get_clean();
     }
 
-    public static function debugOutputNoReset(string $testUrl, string $testMethod = 'get', bool $reset = true): string
+    /**
+     * @throws \Pecee\SimpleRouter\Exceptions\HttpException
+     * @throws NotFoundHttpException
+     * @throws MalformedUrlException
+     */
+    public static function debugOutputNoReset(string $testUrl, string $testMethod = 'get'): string
     {
-        $response = null;
-
         // Route request
         ob_start();
-        static::debugNoReset($testUrl, $testMethod, $reset);
-        $response = ob_get_clean();
 
+        static::debugNoReset($testUrl, $testMethod);
         // Return response
-        return $response;
+        return ob_get_clean();
     }
 
 }
